+++ src/main.rs
@@ -149,7 +149,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
                 .simple_query(
                     "CREATE TYPE avg_ah AS (
                             item_id TEXT,
-                            price DOUBLE PRECISION,
+                            price REAL,
                             sales REAL
                         )",
                 )
